Add UNMANAGED InstallStatus value for components
The UNMANAGED InstallStatus is returned from components operations to indicate
that the component is installed (detected) but is not to be managed by the
Component Manager.  The Component Manager will not take actions (except
checkStatus) until the UNMANAGED status for the component gets resolved
(externally).

The Salt Minion (salt_minion) component returns code 106 ("ExternalInstall")
is equivalent to UNMANAGED.