Improve POSIX guest identification
Many Linux distros do not provide LSB information.  Instead,
they rely exclusively on the the os-release standard information.

We previously used a ranking system that rates the LSB data
(if present) against the os-release data (if present).
With the LSB data install much more rare than before, perform
a quick check for the LSB before attempting expensive operations.

This is in response to these GitHub issues:

https://github.com/vmware/open-vm-tools/issues/647
https://github.com/vmware/open-vm-tools/issues/648

Note that we cannot look just for the os-release data.  In older
Linux releases, there are mistakes or differences between the
os-release and LSB data.  We must continue to do what we do now
or risk changing the identification of older releases.

This optimization is very effective.