Update shutdown code to work for Bottlerocket
Bottlerocket doesn't have /bin/sh which makes the call system() fail
with an unexpected error code. This results in the guest shutdown and
reboot calls via vmtoolsd returning as if the request was successful but
the call actually failed completely. This commit replaces the current
call to system() with g_spawn_sync() which should be a drop in
replacement for system().

Signed-off-by: Matthew Yeazel <yeazelm@amazon.com>