Checking flag 'disable_vmware_customization' in more cloud-init config files
Currently, deployPkg plugin checks the existence of flag
'disable_vmware_customization: false' in the /etc/cloud/cloud.cfg file
to determine if VMware customization is enabled or not on cloud-init
side when cloud-init is available in guest.
Both cloud-init team and customers suggested that it's better practice to
put local configuration like this flag into some .cfg files under
/etc/cloud/cloud.cfg.d directory, ex: /etc/cloud/cloud.cfg.d/somefile.cfg

This change implements the following adjustments to make sure we handle
this flag the same way as cloud-init does in ds-identify and Datasource:
1. Instead of regex matching flag 'disable_vmware_customization: false',
we will check the value of flag 'disable_vmware_customization':
If the value is 'false', it means VMware customization is enabled.
If the value is 'true', it means VMware customization is disabled.
If the flag is not set, by default VMware customization is disabled
on cloud-init side.
2. Besides cloud-init /etc/cloud/cloud.cfg file, we will check all .cfg
files under /etc/cloud/cloud.cfg.d directory.
3. The value of flag 'disable_vmware_customization' in .cfg files under
/etc/cloud/cloud.cfg.d directory will overwrite the one in
/etc/cloud/cloud.cfg file.
4. The value of flag 'disable_vmware_customization' in a .cfg file listed
further down the alphabetical order under /etc/cloud/cloud.cfg.d directory
will overwrite the value in a .cfg file listed earier.
5. If a cloud-init config file contains more than one instance of this
flag, the value of the later flag will overwrite the former one's.

Github Issue: https://github.com/vmware/open-vm-tools/issues/310