File descriptor vulnerability in the open-vm-tools vmware-user-suid-wrapper on Linux
Move the privilege drop logic (dropping privilege to the real uid and
gid of the process for the vmusr service) from suidWrapper to vmtoolsd code.
Now vmtoolsd is not executed with dropped privileges (started as setuid
program) and the dumpable attribute of the process is not reset.
Unprivileged users will not have access to the privileged file descriptors
in the vmtoolsd vmusr process.
Also, set the FD_CLOEXEC flag for both uinputFd and blockFd preventing
those file descriptors from being inherited any further from vmtoolsd.