Change in Guest Data Producer plugin API and Guest Data Producer protocol/handshake between the host and guest to support ‘no subscribers’ error feedback.
The Service Discovery plugin was updated for API compatibility.

Changes in the GDP protocol/handshake introduce versioning of the protocol.
The new version is “2” and introduces new attributes (‘version’ and
‘requireSubs’) in the guest request header sent to the host publisher service.
The response from the host publisher service to the guest producer client is
also versioned and changes more significantly: V2 response messages have new
attributes (‘version’, ‘error-id’, and ‘error-text’) and have the ‘status’
attributes type change from string (‘ok’, ‘bad’) to Boolean (‘true’, ‘false’).
The ‘error-id’, when present, contains a string identifier for an error
type/code and ‘error-text’, when present, contains textual details for the error.

The host publisher supports all protocol versions up to the version it provides
for backward and forward compatibility. The host publisher service responds with
the same version of protocol as the incoming request or its highest protocol
version when the incoming request version is higher than it supports.
The guest producer client supports all response versions up to the version it
provides for backward and forward compatibility. The guest producer client
always sends request to the host publisher service using the highest protocol
version it supports.