Updated NetworkManager calls in suspend/resume scripts.
The org.freedesktop.NetworkManager.Enable method used by the
suspend/resume scripts changes the NetworkEnabled state in NM,
which is disk-persisted.

If the resume script fails for any reason at all, the system will be
left with NetworkManager in a disabled state, where the service
is running, but all interfaces are ignored by NM. This change was made
in commit 13ab4cc0a376993ac8b946a3a1f22126a4a36635 due to NM 0.9
on Ubuntu 11.10 not responding to the Sleep method. I have confirmed
that at least NM 0.9.1 on Ubuntu 11.10 correctly responds to Sleep,
and several modern Linux distributions with modern copies of NM
respond correctly.

Use of the Sleep method is preferred, as the change does not persist a
cold (re)boot, and avoids potentially leaving the system in a confusing
state.