[Coverity]: UNINIT in resolutionSet finding from static application             security testing (SAST)
resolutionSet.c -- 1 issue reported in the file
issue: capabilityArray not initialized or partially initialized when
       reaching the statement.
impact: False-positive
fix: suppress 'uninit_use_in_call'

VMTools_WrapArray converts the capabilityArray to a GArray.
The 'wrapped' array is allocated space for capabilityCount elements and
only the first capabilityCount elements are COPIED from capabilityArray
to the allocated GArray.  As such, the uninitialized elements of the
capabilityArray are never used to generate the returned GArray.  While
technically true (uninitialized) there is no reason to zero the array
indexes from capabilityCount through to the end of the array as these
elements are unused.