vm_support update: gather log files based on the current tools.conf configuration
The script's revised logic replaces the <variable>(s) in the data field
value (logfile name) with a wildcard ('*') character, and it escapes the
'$' sign in the sed match pattern to match the '${\<variable>}' syntax
in the data field.

The usage/help message for the subcommand "get \section> [key]" has also
been updated.