Fix a couple of memory leaks in drag-and-drop/copy-paste code:
* `CopyPasteRpcV3::HandleMsg` calls `CPClipboard_Unserialize`
which allocates memory to the supplied `Clipboard`
structure, but it neglected to call `CPClipboard_Destroy`
to free it when done.  Also update the documentation for
`CPClipboard_Unserialize` to make the contract more
explicit.
* `DnD_SetCPClipboardFromLocalText` completely neglected to
free its destination buffer.

Bonus cleanup:

* Make some pointers `const` (which also allows us to remove
some casts).
* Replace a call to `UNICODE_RELEASE_UTF16` with `free` since
the memory was allocated by lib/dnd with `malloc` and was
not allocated by lib/unicode.