Fix the assembler warning
The Warning "found movsd; assuming movsl was meant" was noticed while
building open-vm-tools with gcc 12.2.0 (on Debian 12.2.0-14).
It was because of the change https://sourceware.org/bugzilla/show_bug.cgi?id=29525
made to the GNU assembler code (GNU Binutils for Debian) 2.39.90.20221231.

Replace the instruction "movsd" with "movsl" to avoid this warning.