Provide a tools.conf setting to disable timesync plugin
This change intends to address a customer issue that requires an
in-guest config to disable timesync without having to reboot VM or
restart. Although we have some ways to implement it from host side,
customer claimed the guest administator may not be the administrator
of vSphere so that an in-guest approach is still needed.

This change adds two config options under "timeSync" group:
- disable-all: This disables all time sync including one-time sync
  and periodic sync.
- disable-periodic: This disables periodic sync only.
  Note we cannot disable one-time sync and keep periodic sync enabled
  as this is not allowed according to the product design.

Addresses Issue: https://github.com/vmware/open-vm-tools/issues/302