Define Posix_Printf/Posix_Fprint for FreeBSD
A previous change to lib/misc/posixPosix.c introduced the
`Posix_Printf` and `Posix_Fprintf` implementations only for non-Solaris,
non-Apple, and non-FreeBSD builds. (Note that this set of
conditions is inconsistent with when `posix.h` aliases them
(and various other functions) to their standard POSIX
counterparts.)

Moving the `Posix_Printf`/`Posix_Fprintf` implementations out
of the `#if` sections so that they're always available.