Update regex for checking 'disable_vmware_customization' flag
The regex check for "disable_vmware_customization" is too strict and
fails when there are inline comments after the flag. As a result, the
cloud-init workflow doesn't start.
This change is relaxing the check to allow inline comments. Please note
that non-capturing group is not supported by POSIX regex API, using
capturing group here.