[vgauth] Retry on EBADF when getting pam entry (getpwuid_r, getpwnam_r)
Below are the changes made as part of this update
o resetting the errno value before calling the getpw*_r functions in
  usercheck.c. Some man pages mention this is required as the errno might
  be left unchanged by the function. We're performing a few retries,
  therefore resetting errno at the start of a retry loop ensures we're
  not 'stuck' on the error.
o implementing retry in impersonateLinux.c which includes the reset of errno.
o logging 'errno' in impersonateLinux.c and usercheck.c where relevant.
o On EBADF error, adding g_thread_yield in the retry loop to add some
  delay to the retry.