Coverity issues fixed.
The change is about the handling of the **result parameter in the wrapper
functions. (UsercheckRetryGetpwnam_r and UsercheckRetryGetpwuid_r)

Issues:
o Dereference after null check (FORWARD_NULL)
  var_deref_model: Passing null pointer result to getpwuid_r, or
  getpwnam_r which dereferences it.

o Dereference before null check (REVERSE_INULL)
  check_after_deref: Null-checking result suggests that it may be
  null, but it has already been dereferenced on all paths leading to the
  check

Add null check for 'result' before dereferencing

The current caller will not have this issue.
We could silence the issue, but we should add an assert for development
and return error 'EINVAL' when '**result' parameter is NULL.

Update caller functions to return VGAUTH_E_INVALID_ARGUMENT when return code
is EINVAL.

The condition used validates that 'result' is NULL while it should
validate '*result' is NULL.
This needs to be fixed; replace "!result" with "*result == NULL" to be
explicit or "!*result" to match other usage in vgauth.