[GTK4] Workaround for G->H RTF Text DnD by returning Plain Text Only
This change introduces a workaround for G->H Drag and Drop operations involving
RTF text in GTK4.

In GTK4, when performing G->H DnD with RTF text, we are unable retrieve the raw
GValue/Byte data from the DnD Clipboard, unlike with the CopyPaste clipboard.
As a result, the original format info is lost. Additionally, the DnD clipboard
always returns a trailing "_" char for these dnd G->H with RTF, which needs to
be removed.

To address these issues, the following changes have been made:
* Always return plain text for G->H RTF DnD
* Remove the trailing "_" if present