Update DnDCPTransportGuestRpc::SendPacket to use GuestRpc_XferWithTimeOut without the GuestRpc_GenericCompletionRoutine callback.
Previously, GuestRpc_GenericCompletionRoutine was responsible for freeing
the buffer after the message was processed, but it could sometimes be called
before GuestRpc finished using the buffer, leading to use-after-free errors.

With this change, GuestRpc takes ownership of the buffer and ensures it is used
safely, and the caller frees it only if GuestRpc_XferWithTimeOut fails.